[proxy] Unify DBusClient and DBusListener
* DBusClient uses ruby-dbus and does not rely on busctl anymore.
